routes: review, comment

+++ data/review.js
@@ -11,6 +11,7 @@ import { gymDataFunctions } from './gym.js'
 
 // get review by review's id, return the review object
 async function get(id) {
+    await validation.checkArgumentsExist(id);
     id = await validation.checkObjectId(id, 'review id')
     const reviewsCollection = await reviewCollection();
     const review = await reviewsCollection.findOne({ _id: new ObjectId(id) });
@@ -35,12 +36,13 @@ async function getAll() {
 
 // get a gym's all  review, return a list of reviewIds (string)
 async function getGymReviews(gymId) {
+    await validation.checkArgumentsExist(gymId);
     gymId = await validation.checkObjectId(gymId, 'gym id')
     const reviewsCollection = await reviewCollection();
     if (!await gymDataFunctions.getByGymId(gymId)) {
-        throw `no user have such id`
+        throw `no gym have such id`
     }
-    const reviewList = await reviewsCollection.find({ gymId: new ObjectId(gymId) }).toArray();
+    const reviewList = await reviewsCollection.find({ gymId: gymId }).toArray();
     // If there are no review for the user, this function will return an empty array
     let gymReview = [];
     for (const review of reviewList) {
@@ -51,13 +53,14 @@ async function getGymReviews(gymId) {
 
 // get a user's all previous review, return a list of reviewIds (string)
 async function getUserReviews(userId) {
+    await validation.checkArgumentsExist(userId);
     userId = await validation.checkObjectId(userId, 'userId')
     const reviewsCollection = await reviewCollection();
     // check is user id exist in database, TBD.
     if (!userDataFunctions.getByUserId(userId)) {
         throw `no user have such id`
     }
-    const reviewList = await reviewsCollection.find({ userId: new ObjectId(userId) }).toArray();
+    const reviewList = await reviewsCollection.find({ userId: userId }).toArray();
 
     // If there are no review for the user, this function will return an empty array
     let userReview = [];
@@ -72,22 +75,33 @@ async function getUserReviews(userId) {
 async function create(
     gymId,
     userId,
-    dataOfReview,
+    dateOfReview,
     content,
     rating) {
+    validation.checkArgumentsExist(gymId, userId, dateOfReview, content, rating)
+    validation.checkNonEmptyStrings(gymId, userId, dateOfReview);
     gymId = await validation.checkObjectId(gymId);
     userId = await validation.checkObjectId(userId);
+    dateOfReview = await validation.checkValidDate(dateOfReview);
 
+    rating = await validation.checkValidRating(rating);
+
+    if (!userDataFunctions.getByUserId(userId)) {
+        throw `no user have such id`;
+    }
+    if (!gymDataFunctions.getByGymId(gymId)) {
+        throw `no gym have such id`;
+    }
+    const gymGet = await gymDataFunctions.getByGymId(gymId);
 
     let newReview = {
         gymId: gymId,
         userId: userId,
-        dataOfReview: dataOfReview,
+        dateOfReview: dateOfReview,
         content: content,
         comments: [],
         rating: rating
     }
-
     const reviewsCollection = await reviewCollection();
     const insertInfo = await reviewsCollection.insertOne(newReview);
     if (!insertInfo || !insertInfo.insertedId) {
@@ -106,28 +120,35 @@ async function create(
     let UpdatedgymReviews = await this.getGymReviews(gymId)
     let updatedGym = await gymDataFunctions.getByGymId(gymId)
     let ratings = []
-    for (review of UpdatedgymReviews) {
+    for (let review of UpdatedgymReviews) {
         ratings.push(review.rating)
     }
+
     ratings.push(rating)
     let total = ratings.reduce((acc, c) => acc + c, 0)
     let grade = ((Math.floor((total / ratings.length) * 10)) / 10)
     updatedGym.reviews = UpdatedgymReviews;
     updatedGym.rating = grade;
 
+
     await gymDataFunctions.update(gymId, updatedGym)
     // finally
     const review = await this.get(newId);
+
     return review;
 }
 
 // remove a review
-
 async function removeReview(id) {
-    id = await validation.checkObjectId(id, 'review id')
+    await validation.checkArgumentsExist(id);
+    id = await validation.checkObjectId(id, 'review id');
+    if (!await this.get(id)) {
+        throw `no review have this id`;
+    }
     //get the user and gym before deletion
-    let userId = await this.get(id).userId;
-    let gymId = await this.get(id).gymId;
+    let review = await this.get(id)
+    let userId = review.userId;
+    let gymId = review.gymId;
     // deletion
     const reviewsCollection = await reviewCollection();
     const deletionInfo = await reviewsCollection.findOneAndDelete({
@@ -138,14 +159,15 @@ async function removeReview(id) {
     }
 
     // user collection remove a review
-    let newReviewsUser = await this.getByUserId(userId);
-    let updatedUser = await userDataFunctions.getByUserId(userId)
+
+    let newReviewsUser = await getUserReviews(userId);
+    let updatedUser = await userDataFunctions.getByUserId(userId);
     updatedUser.reviews = newReviewsUser;
-    await userDataFunctions.update(id, updatedUser)
+    await userDataFunctions.update(userId, updatedUser)
 
     // gym collection remove a review
     let updatedGym = await gymDataFunctions.getByGymId(gymId) // the gym object
-    let newReviewGym = await this.getGymReviews(gymReview); // updated review
+    let newReviewGym = await this.getGymReviews(gymId); // updated review
     updatedGym.reviews = newReviewGym; // update review list
     let allRatings = []
     for (review of newReviewGym) {
@@ -166,19 +188,21 @@ async function removeReview(id) {
 async function updateReviewContent(
     id,
     content,
-    dataOfReview
+    dateOfReview
 ) {
+    await validation.checkArgumentsExist(id, content, dateOfReview);
+    dateOfReview = await validation.checkValidDate(dateOfReview);
     id = await validation.checkObjectId(id, 'review id');
     // how to check content and date?
-
+    const oldReview = await get(id);
     const reviewsCollection = await reviewCollection();
     const updatedInfo = await reviewsCollection.findOneAndUpdate(
         { _id: new ObjectId(id) },
-        { $set: { content: content, dataOfReview: dataOfReview } },
+        { $set: { content: content, dateOfReview: dateOfReview } },
         { returnDocument: 'after' }
     );
     const newReview = await get(id);
-    if (JSON.stringify(newReview) === JSON.stringify(newReview)) {
+    if (JSON.stringify(newReview) === JSON.stringify(oldReview)) {
         throw `there's no real update, everything is the same`;
     }
     if (updatedInfo.lastErrorObject.n === 0) {
@@ -191,15 +215,16 @@ async function updateReviewContent(
 async function updateReviewRating(
     id,
     rating,
-    dataOfReview
+    dateOfReview
 ) {
     id = await validation.checkObjectId(id, 'review id');
     // how to check content and date?
 
     const reviewsCollection = await reviewCollection();
+    const oldReview = await this.get(id);
     const updatedInfo = await reviewsCollection.findOneAndUpdate(
         { _id: new ObjectId(id) },
-        { $set: { rating: rating, dataOfReview: dataOfReview } },
+        { $set: { rating: rating, dateOfReview: dateOfReview } },
         { returnDocument: 'after' }
     );
     const newReview = await get(id);
@@ -209,11 +234,12 @@ async function updateReviewRating(
     if (updatedInfo.lastErrorObject.n === 0) {
         throw 'could not update review successfully';
     }
-
-    let gymId = await this.get(id).gymId; // get the gym of the review
+    let review = await this.get(id);
+    let gymId = await review.gymId; // get the gym of the review
     let updatedgymReviews = await this.getGymReviews(gymId) // get the updated review list
-    let updatedGym = await gymDataFunctions.get(gymId) // get the gym object
+    let updatedGym = await gymDataFunctions.getByGymId(gymId) // get the gym object
     updatedGym.reviews = updatedgymReviews; // update gym review list
+
     let ratings = []
     for (review of updatedgymReviews) {
         ratings.push(review.rating)
@@ -221,6 +247,7 @@ async function updateReviewRating(
     let total = ratings.reduce((acc, c) => acc + c, 0)
     let grade = ((Math.floor((total / ratings.length) * 10)) / 10)
     updatedGym.rating = grade;
+
     await gymDataFunctions.update(gymId, updatedGym)
 
     // finally 
@@ -228,16 +255,19 @@ async function updateReviewRating(
 }
 
 async function updateReviewComment(id, updatedReview) {
-    id = await validation.checkObjectId(id, 'comment id');
+    await validation.checkArgumentsExist(id, updatedReview);
+    id = await validation.checkObjectId(id, 'review id');
+
     const reviewsCollection = await reviewCollection();
     const updatedInfo = await reviewsCollection.findOneAndUpdate(
         { _id: new ObjectId(id) },
-        { $set: { comments: updatedReview.comments } },
+        { $set: updatedReview },
         { returnDocument: 'after' }
     );
     if (updatedInfo.lastErrorObject.n === 0) {
         throw 'could not update review successfully';
     }
+    return await get(id);
 
 }
